refactoring and creating renderItem component
