+++ App.js
@@ -2,29 +2,12 @@ import React, {useState, useRef} from 'react';
 import {Text, View, Dimensions, Image} from 'react-native';
 import Carousel, {Pagination} from 'react-native-snap-carousel';
 
-
+import renderItem from './renderItem'
 import data from './data';
 
 export const SLIDER_WIDTH = Dimensions.get('window').width + 30;
 export const ITEM_WIDTH = Math.round(SLIDER_WIDTH * 0.8);
 
-const renderItem = ({item}) => {
-  return (
-    <View
-      style={{
-        borderWidth: 1,
-        padding: 20,
-        borderRadius: 20,
-        alignItems: 'center',
-        backgroundColor: 'white',
-      }}>
-      <Image source={{uri: item.url}} style={{width: 200, height: 200}} />
-      <Text style={{marginVertical: 10, fontSize: 20, fontWeight: 'bold'}}>
-        {item.name}
-      </Text>
-    </View>
-  );
-};
 
 const App = () => {
   const [index, setIndex] = useState(0);
